Initial attempt at InfoSec Spell File
This is the first draft of an InfoSec spell file. I pulled all the entries in https://csrc.nist.gov/glossary?index=A for the entire alphabet.  
I then sorted it, selected the first word from a line and outputted that to this file. This is intial commit is incomplete as I suspect there will be many issues using it in Libre Office or Word.  
Things to be aware of:
+ I have sorted unique and only first word. As such glossary entries like "Access Control Policy" have been reduced to Access.
+ There are invariably acronyms not present in the NIST glossary that are present in others(RFC1983, SANS, etc.). I will upload more complete lists as I can.
+ To use with Microsoft Word, change the extension to .dic and follow the steps here: https://cybertext.wordpress.com/2018/11/30/word-create-a-custom-dictionary-populated-with-thousands-of-terms/
+ To add a spell file to vim, change the extension to .spl and follow the steps in the first reply here: https://unix.stackexchange.com/questions/85538/how-can-i-create-my-own-spelling-file-for-vim